Fixing the Connection with the Database
For the Db to work it needed other temp files so... i add them into a f
older called 'Db' into the main Directory of the project. Because of th
at some properties Changed in the 'VotantesDbController'. And deleted t
he test message from the 'MainWindow' class.

To do:
The Migration. (Before Continuing)